css-style: styled header and sections
Set the box sizing of all elements to border-box to include padding and
border into the width and height calculations. Styled the header element
by specifying the width, height, padding and text-align. Styled the
sections to be the same width and on the same line using float. added
padding andd margin as well. Flexbox would have been a better solution
but using float was part of the homework requirement. Added a div to
index.html to clear the float.

Signed-off-by: astra9 <thedigitalwhiterabbit@gmail.com>