scroll-feature: scroll feature in the section boxes
Limited the height of the section boxes so that text inside of them
would overflow. I set overflow to hidden. When we hover over the section
then the overflow gets a value of scroll. The cursor becomes row resize.
I also added more text to the title tiles in the mobile view.

Signed-off-by: astra9 <thedigitalwhiterabbit@gmail.com>