style-fix: cleared floating in 3rd section
Added the clear:both property to the 3rd section in the tablet viewport.
This way if the previous two sections were to be floated left and right
and their height set in a way as to overlap with the 3rd section, now
the 3rd section would be pushed to the next line since both sides are
cleared.

Signed-off-by: astra9 <thedigitalwhiterabbit@gmail.com>